refactor: organize client into ui behavior service layers
